Move FFI::Library into a dedicated file
.. for consistency.
