+++ lib/ffi/ffi.rb
@@ -32,6 +32,7 @@ require 'ffi/compat'
 require 'ffi/platform'
 require 'ffi/data_converter'
 require 'ffi/types'
+require 'ffi/library_path'
 require 'ffi/library'
 require 'ffi/errno'
 require 'ffi/abstract_memory'
